requirements.txt
openai
anthropic
python-dotenv
google.generativeai
gradio